+++ clients/kafka-common/src/test/java/org/eclipse/hono/client/kafka/producer/KafkaProducerConfigPropertiesTest.java
@@ -46,14 +46,6 @@ public class KafkaProducerConfigPropertiesTest {
     }
 
     /**
-     * Verifies that trying to set a {@code null} client ID throws a {@link NullPointerException}.
-     */
-    @Test
-    public void testThatClientIdCanNotBeSetToNull() {
-        assertThrows(NullPointerException.class, () -> config.setDefaultClientIdPrefix(null));
-    }
-
-    /**
      * Verifies that properties provided with {@link KafkaProducerConfigProperties#setProducerConfig(Map)} are returned
      * in {@link KafkaProducerConfigProperties#getProducerConfig(String)}.
      */
@@ -97,33 +89,17 @@ public class KafkaProducerConfigPropertiesTest {
     }
 
     /**
-     * Verifies that the client ID set with {@link KafkaProducerConfigProperties#setDefaultClientIdPrefix(String)} is
-     * applied when it is NOT present in the configuration.
+     * Verifies that the client ID set in the {@link KafkaProducerConfigProperties#getProducerConfig(String)}
+     * map conforms to the expected format.
      */
     @Test
-    public void testThatClientIdIsApplied() {
-        final String clientId = "the-client";
+    public void testThatCreatedClientIdConformsToExpectedFormat() {
+        final String componentUid = "hono-adapter-amqp-vertx-7548cc6c66-4qhqh_ed7c6ab9cc27";
 
-        config.setProducerConfig(Map.of());
-        config.setDefaultClientIdPrefix(clientId);
-
-        final Map<String, String> producerConfig = config.getProducerConfig("producerName");
-        assertThat(producerConfig.get("client.id")).startsWith(clientId + "-producerName-");
-    }
-
-    /**
-     * Verifies that the client ID set with {@link KafkaProducerConfigProperties#setDefaultClientIdPrefix(String)} is
-     * NOT applied when it is present in the configuration.
-     */
-    @Test
-    public void testThatClientIdIsNotAppliedIfAlreadyPresent() {
-        final String userProvidedClientId = "custom-client";
-
-        config.setProducerConfig(Map.of("client.id", userProvidedClientId));
-        config.setDefaultClientIdPrefix("other-client");
+        config.setProducerConfig(Map.of("client.id", "configuredId"));
+        config.overrideComponentUidUsedForClientId(componentUid);
 
         final Map<String, String> producerConfig = config.getProducerConfig("producerName");
-        assertThat(producerConfig.get("client.id")).startsWith(userProvidedClientId + "-producerName-");
+        assertThat(producerConfig.get("client.id")).matches("configuredId-producerName-" + componentUid + "_\\d+");
     }
-
 }